debug: user_connect
