+++ tee-worker/omni-executor/pumpx/src/pumpx_api/pumpx_methods/mod.rs
@@ -1,5 +1,1 @@
-pub mod add_wallet;
-pub mod create_market_order_unsigned_tx;
-pub mod get_user_trade_info;
 pub mod user_connect;
-pub mod verify_google_code;
