+++ tee-worker/omni-executor/pumpx/src/pumpx_api/pumpx_types/create_market_order_tx.rs
@@ -31,4 +31,3 @@ pub struct CreateMarketOrderTxResponseData {
 }
 
 pub type CreateMarketOrderTxResponse = ApiResponse<CreateMarketOrderTxResponseData>;
-pub type CreateMarketOrderTxResponse = CreateMarketOrderTxBody;
